Changes made based on the PR code review - Change single character property names - Make sure coupling boundry is not broken - Add nested Location in Event entity - Invert if statements for reduced indetention - Lose authority from EventFormFragment and make delegate for activity finish - Add use case for date and time formats - Make sure that views are not declared inline

+++ app/src/main/java/ftn/com/trippleaproject/domain/BaseModel.java
@@ -5,13 +5,13 @@ import java.io.Serializable;
 
 public abstract class BaseModel implements Serializable {
 
-    private int id;
+    private long id;
 
-    public BaseModel(int id) {
+    public BaseModel(long id) {
         this.id = id;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
